+++ packages/terra-folder-tree/CHANGELOG.md
@@ -2,6 +2,14 @@
 
 ## Unreleased
 
+* Fixed
+  * Fixed an issue where folder tree items could not receive focus after data was updated.
+
+## 1.0.0 - (March 29, 2024)
+
+* Fixed
+  * Fixed screenreader reading the wrong indexes for subfolder items.
+
 ## 1.0.0-alpha.10 - (March 14, 2024)
 
 * Changed